feat: rename touched components
