+++ src/app/SpringBoard/WordMatrix.tsx
@@ -1,4 +1,4 @@
-import Tile from './tile';
+import Tile from './Tile';
 
 interface Props {
   playedWords: string[][];
